Safari-611.1.10/LayoutTests/fast/encoding/meta-in-script.html

<html>
<head>
<script>
<meta charset=koi8-r>
</script>
<meta charset=windows-1255>
</head>
<body>
<pre id="log"></pre>
<script>
function log(message)
{
    document.getElementById("log").innerText += message + "\n";
}

if (window.testRunner)
    testRunner.dumpAsText();

if (document.inputEncoding == "windows-1255")
    log("PASS: " + document.inputEncoding);
else
    log("FAIL: " + document.inputEncoding);
</script>
<p>This test passes if the charset is parsed from the meta tag outside the script.
</body>
</html>
