safari-534.53-branch/LayoutTests/fast/encoding/meta-in-script-expected.txt

CONSOLE MESSAGE: line 4: SyntaxError: Parse error
PASS: windows-1255
This test passes if the charset is parsed from the meta tag outside the script.
